Adding a tone knob
Started adding the parameter for a tone knob, needs a lot of work and to deal with oversampling. It is going to be a low pass filter that cuts off high frequencies to change the tone of the vst. Before it was very static - y and high pitched, so regardless it needed a passive filter, but we decided to let the filter also be adjusted